+++ src/main/java/org/libreoffice/ci/gerrit/buildbot/logic/ProjectControl.java
@@ -30,7 +30,7 @@ public interface ProjectControl {
 	void startGerritJob(Change change, PatchSet patchSet);
 	List<GerritJob> getGerritJobs();
 	GerritJob findJobByRevision(String revision);
-	TbJobDescriptor launchTbJob(Platform platform, Set<String> branch, String box);
+	TbJobDescriptor launchTbJob(Platform platform, Set<String> branch, String box, boolean test);
 	TbJobResult setResultPossible(String ticket, String boxId, TaskStatus status, String logurl);
 	void stop();
 	void start();